Added WTimer
Added custom timer object, and started coding movement for the player object

+++ WLUWGameEngine/WAnimation.h
@@ -8,14 +8,14 @@ class WAnimation
 {
 public:
 	// Constructor
-	WAnimation(WTexture* tex, Vector2 imageCount, float switchTime);
+	WAnimation(WTexture* tex, Vector2 imageCount, Uint32 switchTime);
 
 	// Update animation
-	void update(int row, int colBegin, int colEnd, float deltaTime, bool faceRight);
+	void update(int row, int colBegin, int colEnd, Uint32 deltaTime, bool faceRight);
 
 	// Manage information
 	void setTexture(WTexture* tex);
-	void setSwitchTime(float switchTime);
+	void setSwitchTime(Uint32 switchTime);
 
 	// Fetch texture rectangle
 	SDL_Rect* getClip();
@@ -26,8 +26,8 @@ private:
 	Vector2 imageCount;
 	Vector2 currentImage;
 	
-	float switchTime;
-	float totalTime;
+	Uint32 switchTime;
+	Uint32 totalTime;
 
 	int row;
 	int col;
